+++ Todo.md
@@ -21,3 +21,4 @@
 - ~~결과 출력 클래스 분리~~
 - ~~불필요한 클래스 및 테스트코드 제거~~
 - ~~자동차 경주 우승자 출력~~
+- ~~패키지 구조 변경~~
