+++ src/test/java/racingcar/domain/RacingGameTest.java
@@ -1,7 +1,6 @@
-package racingcar;
+package racingcar.domain;
 
 import static org.assertj.core.api.Assertions.*;
-import static org.junit.jupiter.api.Assertions.*;
 
 import java.util.List;
